Chase Demo by Chemdoc77
The following code contains chase animation both for symmetrical and
unsymmetrical chase animations.  The user can chose a single color, a
color from an array of colors or a series of colors from a palette for
the color of the pixels in the color chase bar.  The user can chose the
number of pixels in the color chase bar.

Look at the READ_ME.h tab for further explanation of this sketch.